Let's make this run faster

+++ grabfile.py
@@ -1,24 +1,32 @@
 """For anything concerning grabbing and cleaning up a file."""
 import re
+from pathlib import Path
 
 
 def grab_file():
     """Grab file and splits it all into a list of words."""
     # Please use a relative path if uploading online!
-    filename = "theroom.txt"
-    # Placing individual words into file_input list
+    pregen_histofile = Path("theroomcleanedup.txt")
+    if pregen_histofile.is_file():
+        print("It's already here!")
+        file_open = open(pregen_histofile, encoding='utf8').read()
+        file_input = file_open.split(" ")
+    else:
+        filename = "theroom.txt"
+        # Placing individual words into file_input list
 
-    file_open = open(filename, encoding='utf8').read()
+        file_open = open(filename, encoding='utf8').read()
 
-    file_input = room_service(file_open)
+        file_input = room_service(file_open)
 
-    histofile = open("theroomcleanedup.txt", "w+")
-    histofile.write(' '.join(file_input))
-    histofile.close()
+        histofile = open("theroomcleanedup.txt", "w+")
+        histofile.write(' '.join(file_input))
+        histofile.close()
     # file_input = clean_up(file_input)
     return file_input
 
 def room_service(filename):
+
     """Cleans 'The Room' script."""
     # removes all actions
     actions_removed = re.sub(r"( \(.*\))", '', filename)
@@ -31,6 +39,7 @@ def room_service(filename):
     # only puts in dialogue
     dialogue = re.findall(r"(?:\: (.*))", stop_tokens_quest)
     regular_text = re.findall(r"[\[\]'\-\w\,]+", ' '.join(dialogue).lower())
+    regular_text.insert(0, '[start]')
 
     return(regular_text)
 
@@ -42,5 +51,3 @@ def capitalize_check(text):
         if word in capitalize_these:
             text[value] = word.capitalize()
     return text
-
-grab_file()
